in Ruby 2.0.0, Find.find has an annoying return type
Apparently, in Ruby 2.0.0, Find.find includes the exact argument passed to it as the first element in the returned array, but then, regardless of the passed-in argument's type, the rest of the array
is String instances. In later versions, the entire array is String instances. I could see returning all Pathnames or all Strings, but the first being variable, then everything else always being
strings is just annoying.